[kruskal] changing result columns order

+++ sql/mst/kruskalDD.sql
@@ -30,7 +30,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 CREATE OR REPLACE FUNCTION pgr_kruskalDD (
     TEXT,   -- Edge sql
     BIGINT, -- root vertex
-    FLOAT,  -- distance
+    NUMERIC,  -- distance
 
     OUT seq BIGINT,
     OUT depth BIGINT,
@@ -49,7 +49,7 @@ BEGIN
 
     RETURN QUERY
     SELECT *
-    FROM _pgr_kruskal(_pgr_get_statement($1), ARRAY[$2]::BIGINT[], 'DFS', -1, $3);
+    FROM _pgr_kruskal(_pgr_get_statement($1), ARRAY[$2]::BIGINT[], 'DFS', -1, $3::FLOAT);
 END;
 $BODY$
 LANGUAGE plpgsql VOLATILE STRICT;
@@ -59,7 +59,7 @@ CREATE OR REPLACE FUNCTION pgr_kruskalDD (
     TEXT,   -- Edge sql
     ANYARRAY, -- root vertex
 
-    FLOAT, -- distance
+    NUMERIC, -- distance
 
     OUT seq BIGINT,
     OUT depth BIGINT,
@@ -78,10 +78,10 @@ BEGIN
 
     RETURN QUERY
     SELECT *
-    FROM _pgr_kruskal(_pgr_get_statement($1), $2, 'DFS', -1, $3);
+    FROM _pgr_kruskal(_pgr_get_statement($1), $2, 'DFS', -1, $3::FLOAT);
 END;
 $BODY$
 LANGUAGE plpgsql VOLATILE STRICT;
 
-COMMENT ON FUNCTION pgr_kruskalDD(TEXT, BIGINT, FLOAT) IS 'pgr_kruskalDD(Single vertex): Experimental, Undirected Graph';
-COMMENT ON FUNCTION pgr_kruskalDD(TEXT, ANYARRAY, FLOAT) IS 'pgr_kruskalDD(Multiple vertices): Experimental, Undirected Graph';
+COMMENT ON FUNCTION pgr_kruskalDD(TEXT, BIGINT, NUMERIC) IS 'pgr_kruskalDD(Single vertex): Experimental, Undirected Graph';
+COMMENT ON FUNCTION pgr_kruskalDD(TEXT, ANYARRAY, NUMERIC) IS 'pgr_kruskalDD(Multiple vertices): Experimental, Undirected Graph';
